make driveTime used encoders for straight line

+++ helpers.h
@@ -3,12 +3,12 @@
 
 #include "config.h"
 
-void driveTime(float speed, int millis);
-void driveDistance(float speed, float distance);
-void leftMotor(float speed);
-void rightMotor(float speed);
-void moveMotor(tMotor mot, float speed);
-void pivot(float speed, int millis);
+void driveTime(int speed, int millis);
+void driveDistance(int speed, float distance);
+void leftMotor(int speed);
+void rightMotor(int speed);
+void moveMotor(tMotor mot, int speed);
+void pivot(int speed, int millis);
 void servoAngle(TServoIndex srv, int angle, int max);
 
 #endif
